Move PomFileHelper to util package
